Add Boot and yaml to tests
